Cambios en la base de datos
* Se cambio el tamaño y la precisión de la variable explicó
* Se agregaron valores por defecto a algunas tablas
* Se agrego el poder_maximo_posible

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -18,7 +18,7 @@ class CreateUsersTable extends Migration
             $table->string('uid_user')->unique();
             //$table->string('name');
             $table->string('uid_avatar');
-            $table->integer('exp');
+            $table->decimal('exp',42,2);
             // $table->string('email')->unique();
             // $table->timestamp('email_verified_at')->nullable();
             // $table->string('password');
