+++ database/migrations/2018_10_09_192202_create_victimas_table.php
@@ -17,7 +17,7 @@ class CreateVictimasTable extends Migration
             $table->increments('id');
             $table->string('uid_user');
             $table->string('uid_avatar');
-            $table->integer('exp');
+            $table->decimal('exp',42,2);
             $table->timestamps();
         });
     }
