+++ database/migrations/2018_10_27_220729_create_inscritos_torneos_table.php
@@ -17,7 +17,7 @@ class CreateInscritosTorneosTable extends Migration
             $table->string('uid_user');
             $table->string('uid_torneo');
             $table->string('uid_avatar');
-            $table->integer('exp');
+            $table->decimal('exp',42,2);
             $table->timestamps();
         });
     }
